Fixing i3 tests that were failing

+++ core/src/main/java/com/thimbleware/jmemcached/CacheElement.java
@@ -15,9 +15,9 @@ public interface CacheElement extends Serializable, SizedItem {
 
     int hashCode();
 
-    long getExpire();
+    long getExpire(); // uint
     
-    void setExpire(long expire);
+    void setExpire(long expire); // uint
 
     int getFlags();
 
@@ -41,5 +41,5 @@ public interface CacheElement extends Serializable, SizedItem {
 
     CacheElement prepend(LocalCacheElement element);
 
-    LocalCacheElement.IncrDecrResult add(int mod);
+    LocalCacheElement.IncrDecrResult add(long mod);
 }
